concept variable prediction, even for nested rules
Summary:
- sidenote: recursion ftw
- actual predictions for a concept variable
- to do this, we have to modify the set of rules to feed to
applyDetector (our inference function)
- also, fixed case where a concept variable with nested concepts doesn't
find the proper yelp categories to do the simulation on
- Refactored core detector code parsing into the lib/detectors/detector.js file
