+++ client/templates/blockly.js
@@ -77,115 +77,6 @@ export class ReflectAndExpand {
 
 }
 
-export const splitVarDeclarationAndRules = function(code) {
-  let lines_w_comment = code.split('\n');
-  let lines = filterFunctionalLines(lines_w_comment);
-  let rules = lines.slice(1, lines.length);
-  // let singleLineVarDecl = lines[0];
-  let varDecl = lines[0];
-  return [varDecl, rules];
-};
-
-const filterFunctionalLines = (lines_w_comment) => {
-  // filter to functional code lines
-  let lines = lines_w_comment.filter((line) => {
-    return ((line.slice(0,2) != "//" ) && // remove comments
-            (line != "")) // remove new lines
-  });
-  return lines
-}
-
-export const getConceptVariables = (rules) => {
-  console.log("rules: ", rules);
-  if (!Array.isArray(rules)) {
-    console.log("rules is not an array");
-    return [[], []];
-  }
-
-  if (rules.length < 1) {
-    console.log("rules is empty array");
-    return [[], []];
-  }
-
-  let conceptVariableNames = [];
-  let conceptVariableRules = [];
-  rules.forEach(function(rule) {
-    let split_rule = rule.split("=");
-
-    // a concept variable rule looks like
-    // "concept_variable = feature1 || feature2"
-    // it follows this format
-    if (split_rule.length == 2) {
-      let conceptVariableName = split_rule[0].trim();
-      let conceptRule = split_rule[1];
-      conceptVariableNames.push(conceptVariableName);
-      conceptVariableRules.push(conceptRule);
-    }
-  });
-  return [conceptVariableNames, conceptVariableRules];
-}
-
-/**
- *
- * @param {*} variableNames List of Strings
- *   obtained from a call like `variableNames = setOfVariableNames(varDecl);`
- * @param {*} rule String
- *     "concept_variable = feature1 || feature2"
- *     "feature1 || (feature2 && feature3)"
- *     "(feature1 || (feature2 && !(feature3)))"
- * @returns {*} context_features List of Strings
- *
- */
-export const conceptRulesToContextFeatures = (variableNames, rule) => {
-  return variableNames.filter((name) => { return rule.includes(name) });
-}
-
-/* must read the full set of context features in the blockly workspace */
-export const setOfContextFeaturesInBlockly = (varDecl, rules) => {
-  let variableNames = setOfVariableNames(varDecl);
-  console.log("variableNames: ", variableNames);
-
-  let [conceptVariableNames, conceptVariableRules] = getConceptVariables(rules);
-  console.log("concept variables: ", conceptVariableNames);
-
-  let contextFeatures = variableNames.filter((variable) => {
-    return !conceptVariableNames.includes(variable)
-  });
-
-  console.log("contextFeatures: ", contextFeatures);
-  return contextFeatures;
-}
-
-export const setOfVariableNames = (varDecl) => {
-  console.log("varDecl: ", varDecl);
-  if (varDecl === undefined) {
-    return [];
-  }
-
-  if (typeof(varDecl, String) && (varDecl.length > 0)) {
-    let varDeclClean = varDecl.slice(
-        4, // skip "var "
-        varDecl.length - 1 // skip ending ";"
-      );
-    let variablesSplit = varDeclClean.split(', ');
-    console.log('variablesSplit: ', variablesSplit);
-    return variablesSplit;
-    // if (variablesSplit.length == 1) {
-      // return variablesSplit;
-    // }
-    // return variablesSplit.map(variable => { variable.trim() });
-  }
-  return [];
-}
-
-export const vardecl2placecategory = (var_placecat) => {
-  // "var beaches;"
-  let placecat = var_placecat.split(' ')[1];
-  // "beaches;" or "beaches"
-  placecat = placecat.slice(-1) == ';' ? placecat.slice(0,-1) : placecat;
-  return placecat;
-};
-
 export const mockTestDetector = function (userAffordances, varDecl, rules) {
   // userAffordances: key value pairs of (elementaryContext: values)
   // varDecl: list of strings of JS let declaration
